refactor(dns): share is_private_ip in wardnet-common; cover mapped-v6 + CGNAT (#3)
Promote IP classification to a single wardnet_common::net module used by
both the DDNS SSRF guard (is_reserved_ipv4) and DNS rebinding protection
(is_private_ip), replacing two divergent copies. Broaden coverage so
rebinding can't be bypassed:
- IPv4-mapped IPv6 (::ffff:192.168.1.1) — classified via the embedded v4.
- RFC 6598 CGNAT 100.64.0.0/10, broadcast, documentation (already in the
  DDNS guard, now also applied to rebinding).
Deletes wardnetd/src/dns/rebinding.rs (logic moved + extended).
