+++ source/daemon/crates/wardnetd/src/dns/mod.rs
@@ -1,7 +1,6 @@
 pub mod server;
 
 mod rate_limit;
-mod rebinding;
 
 #[cfg(test)]
 mod tests;
